Give better error messages when dependencies are not present
This way in the common case where a dependency is not present we will
give a better error than "attribute missing".

This will bloat the size of the generated files a bit since a lot of the
messages are repeated, but I couldn't see a nice way to do it otherwise
while keeping the `ToNixExpr` functions self-contained.

I was a little bit unsure about `ExeDependency` - I *think* it's trying
to look in `components.exes` although at the moment it just looks at the
ambient scope. I made this explicit.