Fix @rev:<n> extra-deps specification not working
This mistake was introduced in e04dbd4ea7cb7eeb00acdbe4d4e1b7b66dcb19d7.
Where previously the (+++) operator from
https://hackage.haskell.org/package/Cabal-2.4.1.0/docs/Distribution-Compat-ReadP.html#v:-43--43--43-
was used to give no precedence to left or right,
now the <|> operator is used instead, which fails if the left fails and
won't try the right (non-backtracking).

To fix this, a try is used,
which explicitly makes the code backtrack if the left was unsuccessful.

A better fix would be possible by reading until ':', then change
behavior depending on whether "sha256" or "rev" was read